+++ packages/open-workflow-diagram-editor/stories/features/UndoRedo.stories.tsx
@@ -14,14 +14,499 @@
  * limitations under the License.
  */
 
+import React, { useCallback, useEffect, useRef, useState } from "react";
 import type { Meta, StoryObj } from "@storybook/react-vite";
 import { useArgs } from "storybook/preview-api";
-import { UndoRedoEditor } from "./UndoRedoEditor";
+import { DiagramEditor, DiagramEditorRef } from "../../src/diagram-editor/DiagramEditor";
+import { useResolvedColorMode } from "../../src/hooks/useResolvedColorMode";
 import { authenticationReusable } from "../examples";
+
+// ---------------------------------------------------------------------------
+// Toolbar theme tokens
+// ---------------------------------------------------------------------------
+
+type Theme = {
+  toolbar: React.CSSProperties;
+  button: React.CSSProperties;
+  buttonPressed: { background: string; boxShadow: string };
+};
+
+const light: Theme = {
+  toolbar: { borderBottom: "1px solid #e5e7eb", background: "#f7f8fa" },
+  button: {
+    border: "1px solid #d1d5db",
+    background: "linear-gradient(to bottom, #ffffff, #f3f4f6)",
+    color: "#374151",
+    boxShadow: "0 1px 2px rgba(0,0,0,0.08), inset 0 1px 0 rgba(255,255,255,0.9)",
+  },
+  buttonPressed: {
+    background: "linear-gradient(to bottom, #e5e7eb, #f3f4f6)",
+    boxShadow: "0 0 0 rgba(0,0,0,0), inset 0 1px 3px rgba(0,0,0,0.15)",
+  },
+};
+
+const dark: Theme = {
+  toolbar: { borderBottom: "1px solid #374151", background: "#1f2937" },
+  button: {
+    border: "1px solid #4b5563",
+    background: "linear-gradient(to bottom, #374151, #2d3748)",
+    color: "#e5e7eb",
+    boxShadow: "0 1px 2px rgba(0,0,0,0.4), inset 0 1px 0 rgba(255,255,255,0.06)",
+  },
+  buttonPressed: {
+    background: "linear-gradient(to bottom, #1f2937, #2d3748)",
+    boxShadow: "0 0 0 rgba(0,0,0,0), inset 0 2px 4px rgba(0,0,0,0.4)",
+  },
+};
+
+const baseButtonStyle: React.CSSProperties = {
+  display: "inline-flex",
+  alignItems: "center",
+  gap: "6px",
+  padding: "4px 12px",
+  height: "32px",
+  fontSize: "13px",
+  fontFamily: "inherit",
+  fontWeight: 500,
+  lineHeight: 1,
+  cursor: "pointer",
+  borderRadius: "6px",
+  userSelect: "none",
+  transition: "background 60ms, box-shadow 60ms, transform 60ms",
+};
+
+// ---------------------------------------------------------------------------
+// Story render function
+// ---------------------------------------------------------------------------
+
+function UndoRedoStory({
+  content,
+  colorMode: colorModeProp,
+  isReadOnly,
+  locale,
+  onContentChange,
+}: {
+  content: string;
+  colorMode?: string;
+  isReadOnly?: boolean;
+  locale?: string;
+  onContentChange?: (content: string) => void;
+}) {
+  const editorRef = useRef<DiagramEditorRef | null>(null);
+  const resolvedColorMode = useResolvedColorMode(
+    (colorModeProp as "light" | "dark" | "system") ?? "system",
+  );
+  const theme = resolvedColorMode === "dark" ? dark : light;
+
+  const [canUndo, setCanUndo] = useState(false);
+  const [canRedo, setCanRedo] = useState(false);
+  const [modalOpen, setModalOpen] = useState(false);
+  const [modalText, setModalText] = useState("");
+  const [getContentOpen, setGetContentOpen] = useState(false);
+  const [getContentText, setGetContentText] = useState("");
+  const [copied, setCopied] = useState(false);
+
+  // True only while a button-triggered undo/redo is pending its deferred sync.
+  // Prevents onContentChange from firing when content changes externally
+  // (e.g. via the Controls panel), which would create a feedback loop.
+  const undoRedoInFlight = useRef(false);
+
+  const syncHistory = useCallback(() => {
+    setCanUndo(editorRef.current?.canUndo ?? false);
+    setCanRedo(editorRef.current?.canRedo ?? false);
+    if (undoRedoInFlight.current && onContentChange && editorRef.current) {
+      onContentChange(editorRef.current.getContent());
+    }
+    undoRedoInFlight.current = false;
+  }, [onContentChange]);
+
+  // Sync button state after external content changes — never notify the host.
+  useEffect(() => {
+    const id = setTimeout(syncHistory, 0);
+    return () => clearTimeout(id);
+  }, [content, syncHistory]);
+
+  // Expose the ref on the window for browser-console testing.
+  useEffect(() => {
+    (window as unknown as Record<string, unknown>).diagramEditorRef = editorRef;
+    return () => {
+      delete (window as unknown as Record<string, unknown>).diagramEditorRef;
+    };
+  }, []);
+
+  // ---------------------------------------------------------------------------
+  // Button helpers
+  // ---------------------------------------------------------------------------
+
+  const buttonStyle = (disabled: boolean): React.CSSProperties => ({
+    ...baseButtonStyle,
+    ...theme.button,
+    ...(disabled ? { opacity: 0.4, cursor: "not-allowed", pointerEvents: "none" } : {}),
+  });
+
+  const handlePointerDown = (e: React.PointerEvent<HTMLButtonElement>) => {
+    const btn = e.currentTarget;
+    btn.style.background = theme.buttonPressed.background;
+    btn.style.boxShadow = theme.buttonPressed.boxShadow;
+    btn.style.transform = "translateY(1px)";
+  };
+
+  const handlePointerUp = (e: React.PointerEvent<HTMLButtonElement>) => {
+    const btn = e.currentTarget;
+    btn.style.background = theme.button.background as string;
+    btn.style.boxShadow = theme.button.boxShadow as string;
+    btn.style.transform = "";
+  };
+
+  // ---------------------------------------------------------------------------
+  // Set Content modal
+  // ---------------------------------------------------------------------------
+
+  const openSetContentModal = () => {
+    setModalText(editorRef.current?.getContent() ?? "");
+    setModalOpen(true);
+  };
+
+  const applyModal = () => {
+    editorRef.current?.setContent(modalText);
+    setModalOpen(false);
+    setTimeout(syncHistory, 0);
+  };
+
+  // ---------------------------------------------------------------------------
+  // Get Content modal
+  // ---------------------------------------------------------------------------
+
+  const openGetContentModal = () => {
+    setGetContentText(editorRef.current?.getContent() ?? "");
+    setCopied(false);
+    setGetContentOpen(true);
+  };
+
+  const copyToClipboard = () => {
+    navigator.clipboard.writeText(getContentText).then(() => {
+      setCopied(true);
+      setTimeout(() => setCopied(false), 2000);
+    });
+  };
+
+  // ---------------------------------------------------------------------------
+  // Shared modal styles
+  // ---------------------------------------------------------------------------
+
+  const overlayStyle: React.CSSProperties = {
+    position: "fixed",
+    inset: 0,
+    background: "rgba(0,0,0,0.45)",
+    display: "flex",
+    alignItems: "center",
+    justifyContent: "center",
+    zIndex: 1000,
+  };
+
+  const dialogStyle: React.CSSProperties = {
+    display: "flex",
+    flexDirection: "column",
+    width: "min(660px, 92vw)",
+    maxHeight: "80vh",
+    borderRadius: "10px",
+    overflow: "hidden",
+    boxShadow: "0 20px 60px rgba(0,0,0,0.3)",
+    background: resolvedColorMode === "dark" ? "#1f2937" : "#ffffff",
+    border: resolvedColorMode === "dark" ? "1px solid #374151" : "1px solid #d1d5db",
+  };
+
+  const dialogHeaderStyle: React.CSSProperties = {
+    padding: "14px 20px",
+    fontWeight: 600,
+    fontSize: "15px",
+    borderBottom: resolvedColorMode === "dark" ? "1px solid #374151" : "1px solid #e5e7eb",
+    color: resolvedColorMode === "dark" ? "#f3f4f6" : "#111827",
+    flexShrink: 0,
+  };
+
+  const textareaStyle: React.CSSProperties = {
+    flex: 1,
+    resize: "none",
+    border: "none",
+    outline: "none",
+    padding: "16px 20px",
+    fontFamily: '"Menlo", "Consolas", "Monaco", monospace',
+    fontSize: "12.5px",
+    lineHeight: 1.6,
+    background: resolvedColorMode === "dark" ? "#111827" : "#f7f8fa",
+    color: resolvedColorMode === "dark" ? "#e5e7eb" : "#1f2328",
+    overflowY: "auto",
+    minHeight: "320px",
+  };
+
+  const dialogFooterStyle: React.CSSProperties = {
+    display: "flex",
+    justifyContent: "flex-end",
+    gap: "8px",
+    padding: "12px 16px",
+    borderTop: resolvedColorMode === "dark" ? "1px solid #374151" : "1px solid #e5e7eb",
+    flexShrink: 0,
+  };
+
+  // ---------------------------------------------------------------------------
+  // Render
+  // ---------------------------------------------------------------------------
+
+  return (
+    <>
+      {/* Get Content modal */}
+      {getContentOpen && (
+        <div style={overlayStyle} onMouseDown={() => setGetContentOpen(false)}>
+          <div style={dialogStyle} onMouseDown={(e) => e.stopPropagation()}>
+            <div style={dialogHeaderStyle}>Get Content</div>
+            <textarea
+              style={{ ...textareaStyle, cursor: "default", userSelect: "text" }}
+              value={getContentText}
+              readOnly
+              spellCheck={false}
+            />
+            <div style={dialogFooterStyle}>
+              <button
+                style={buttonStyle(false)}
+                onClick={copyToClipboard}
+                onPointerDown={handlePointerDown}
+                onPointerUp={handlePointerUp}
+                onPointerLeave={handlePointerUp}
+              >
+                <svg width="13" height="13" viewBox="0 0 16 16" fill="none" aria-hidden="true">
+                  <rect
+                    x="5"
+                    y="5"
+                    width="9"
+                    height="10"
+                    rx="1.5"
+                    stroke="currentColor"
+                    strokeWidth="1.5"
+                  />
+                  <path
+                    d="M11 5V3.5A1.5 1.5 0 0 0 9.5 2h-7A1.5 1.5 0 0 0 1 3.5v7A1.5 1.5 0 0 0 2.5 12H4"
+                    stroke="currentColor"
+                    strokeWidth="1.5"
+                    strokeLinecap="round"
+                  />
+                </svg>
+                {copied ? "Copied!" : "Copy to Clipboard"}
+              </button>
+              <button
+                style={buttonStyle(false)}
+                onClick={() => setGetContentOpen(false)}
+                onPointerDown={handlePointerDown}
+                onPointerUp={handlePointerUp}
+                onPointerLeave={handlePointerUp}
+              >
+                Close
+              </button>
+            </div>
+          </div>
+        </div>
+      )}
+
+      {/* Set Content modal */}
+      {modalOpen && (
+        <div style={overlayStyle} onMouseDown={() => setModalOpen(false)}>
+          <div style={dialogStyle} onMouseDown={(e) => e.stopPropagation()}>
+            <div style={dialogHeaderStyle}>Set Content</div>
+            <textarea
+              style={textareaStyle}
+              value={modalText}
+              onChange={(e) => setModalText(e.target.value)}
+              spellCheck={false}
+              autoFocus
+            />
+            <div style={dialogFooterStyle}>
+              <button
+                style={buttonStyle(false)}
+                onClick={() => setModalOpen(false)}
+                onPointerDown={handlePointerDown}
+                onPointerUp={handlePointerUp}
+                onPointerLeave={handlePointerUp}
+              >
+                Cancel
+              </button>
+              <button
+                style={{
+                  ...buttonStyle(false),
+                  background: "#3b82d4",
+                  color: "#fff",
+                  border: "1px solid #2563eb",
+                }}
+                onClick={applyModal}
+                onPointerDown={handlePointerDown}
+                onPointerUp={handlePointerUp}
+                onPointerLeave={handlePointerUp}
+              >
+                Apply
+              </button>
+            </div>
+          </div>
+        </div>
+      )}
+
+      <div style={{ display: "flex", flexDirection: "column", height: "100vh" }}>
+        {/* Toolbar — exercises the DiagramEditorRef API */}
+        <div
+          style={{
+            display: "flex",
+            gap: "8px",
+            padding: "6px 12px",
+            flexShrink: 0,
+            alignItems: "center",
+            ...theme.toolbar,
+          }}
+        >
+          <button
+            style={buttonStyle(!canUndo)}
+            disabled={!canUndo}
+            onClick={() => {
+              undoRedoInFlight.current = true;
+              editorRef.current?.undo();
+              setTimeout(syncHistory, 0);
+            }}
+            onPointerDown={handlePointerDown}
+            onPointerUp={handlePointerUp}
+            onPointerLeave={handlePointerUp}
+            aria-label="Undo"
+          >
+            <svg width="14" height="14" viewBox="0 0 16 16" fill="none" aria-hidden="true">
+              <path
+                d="M3.5 6H9a4 4 0 0 1 0 8H5"
+                stroke="currentColor"
+                strokeWidth="1.5"
+                strokeLinecap="round"
+                strokeLinejoin="round"
+              />
+              <path
+                d="M3.5 3.5 1 6l2.5 2.5"
+                stroke="currentColor"
+                strokeWidth="1.5"
+                strokeLinecap="round"
+                strokeLinejoin="round"
+              />
+            </svg>
+            Undo
+          </button>
+          <button
+            style={buttonStyle(!canRedo)}
+            disabled={!canRedo}
+            onClick={() => {
+              undoRedoInFlight.current = true;
+              editorRef.current?.redo();
+              setTimeout(syncHistory, 0);
+            }}
+            onPointerDown={handlePointerDown}
+            onPointerUp={handlePointerUp}
+            onPointerLeave={handlePointerUp}
+            aria-label="Redo"
+          >
+            Redo
+            <svg width="14" height="14" viewBox="0 0 16 16" fill="none" aria-hidden="true">
+              <path
+                d="M12.5 6H7a4 4 0 0 0 0 8h4"
+                stroke="currentColor"
+                strokeWidth="1.5"
+                strokeLinecap="round"
+                strokeLinejoin="round"
+              />
+              <path
+                d="M12.5 3.5 15 6l-2.5 2.5"
+                stroke="currentColor"
+                strokeWidth="1.5"
+                strokeLinecap="round"
+                strokeLinejoin="round"
+              />
+            </svg>
+          </button>
+          <button
+            style={buttonStyle(false)}
+            onClick={openSetContentModal}
+            onPointerDown={handlePointerDown}
+            onPointerUp={handlePointerUp}
+            onPointerLeave={handlePointerUp}
+            aria-label="Set Content"
+          >
+            <svg width="14" height="14" viewBox="0 0 16 16" fill="none" aria-hidden="true">
+              <rect
+                x="2"
+                y="2"
+                width="12"
+                height="12"
+                rx="2"
+                stroke="currentColor"
+                strokeWidth="1.5"
+              />
+              <path
+                d="M5 6h6M5 9h4"
+                stroke="currentColor"
+                strokeWidth="1.5"
+                strokeLinecap="round"
+              />
+            </svg>
+            Set Content
+          </button>
+          <button
+            style={buttonStyle(false)}
+            onClick={openGetContentModal}
+            onPointerDown={handlePointerDown}
+            onPointerUp={handlePointerUp}
+            onPointerLeave={handlePointerUp}
+            aria-label="Get Content"
+          >
+            <svg width="14" height="14" viewBox="0 0 16 16" fill="none" aria-hidden="true">
+              <rect
+                x="2"
+                y="2"
+                width="12"
+                height="12"
+                rx="2"
+                stroke="currentColor"
+                strokeWidth="1.5"
+              />
+              <path
+                d="M5 6h6M5 9h3"
+                stroke="currentColor"
+                strokeWidth="1.5"
+                strokeLinecap="round"
+              />
+              <path
+                d="M11 10l2 2-2 2"
+                stroke="currentColor"
+                strokeWidth="1.5"
+                strokeLinecap="round"
+                strokeLinejoin="round"
+              />
+            </svg>
+            Get Content
+          </button>
+        </div>
+
+        {/* DiagramEditor — fills remaining height */}
+        <div style={{ flex: 1, overflow: "hidden" }}>
+          <DiagramEditor
+            ref={editorRef}
+            content={content}
+            isReadOnly={isReadOnly}
+            locale={locale}
+            colorMode={colorModeProp as "light" | "dark" | "system" | undefined}
+          />
+        </div>
+      </div>
+    </>
+  );
+}
+
+// ---------------------------------------------------------------------------
+// Storybook meta
+// ---------------------------------------------------------------------------
 
 const meta = {
   title: "Features/Undo Redo",
-  component: UndoRedoEditor,
+  component: DiagramEditor,
   tags: ["autodocs"],
   parameters: {
     layout: "fullscreen",
@@ -148,8 +633,8 @@ history stack.
 ### Syncing \`canUndo\` / \`canRedo\` into React state
 
 \`canUndo\` and \`canRedo\` are **plain values on the ref object**, not reactive
-state. To drive toolbar buttons, copy them into local state after each
-operation and after external content changes:
+state. To drive UI controls such as toolbar buttons, copy them into local state
+after each operation and after external content changes:
 
 \`\`\`tsx
 const [canUndo, setCanUndo] = useState(false);
@@ -181,6 +666,17 @@ const handleSetContent = (newContent: string) => {
 
 > The \`setTimeout(..., 0)\` deferral is necessary because \`useImperativeHandle\`
 > updates the ref values one render after the internal state changes.
+
+---
+
+### Story toolbar
+
+The toolbar rendered above the diagram in this story (**Undo**, **Redo**,
+**Set Content**, **Get Content**) is implemented entirely in the story's render
+function — it is **not** part of the \`DiagramEditor\` component. Its purpose is
+to provide an interactive way to exercise the ref API and to show one approach
+to wiring up a toolbar in a host application. You are free to implement your
+own toolbar or menu in whatever way suits your application.
 
 ---
 
@@ -222,9 +718,8 @@ do:
 \`);
 \`\`\`
 
-> **Tip:** The toolbar buttons above the diagram (**Undo**, **Redo**,
-> **Set Content**, **Get Content**) call the same ref methods, so you can
-> freely mix toolbar interactions with console calls.
+> **Tip:** The toolbar buttons above the diagram call the same ref methods, so
+> you can freely mix toolbar interactions with console calls.
         `,
       },
     },
@@ -232,15 +727,18 @@ do:
   render: (args, { globals }) => {
     // eslint-disable-next-line react-hooks/rules-of-hooks
     const [, updateArgs] = useArgs();
+    const colorMode = args.colorMode ?? globals.colorMode ?? "system";
     return (
-      <UndoRedoEditor
-        {...args}
-        colorMode={args.colorMode ?? globals.colorMode ?? "system"}
+      <UndoRedoStory
+        content={args.content ?? ""}
+        colorMode={colorMode}
+        isReadOnly={args.isReadOnly}
+        locale={args.locale}
         onContentChange={(content) => updateArgs({ content })}
       />
     );
   },
-} satisfies Meta<typeof UndoRedoEditor>;
+} satisfies Meta<typeof DiagramEditor>;
 
 export default meta;
 type Story = StoryObj<typeof meta>;
